refactor(web): Components 폴더 내부 컴포넌트들 파일 구조 변경 (#118)
* feat(web): 포스팅 닉네임 클릭시 마이페이지 이동 처리

* design(web): 마이페이지 패딩처리

* refactor(web): 컴포넌트 내부 컴포넌트가 여러개일 때 내부 components 폴더로 분리

* fix(web): toast-ui editor not ssr 해결
