+++ apps/web/src/components/account/profile/components/index.ts
@@ -1,0 +1,3 @@
+export { default as ProfileInfo } from './profileInfo';
+export { default as ProfileRepresentative } from './profileRepresentative';
+export { default as ProfileBackground } from './profileBackground';
