+++ apps/web/src/components/common/comment/components/index.ts
@@ -1,0 +1,3 @@
+export { default as CommentArea } from './commentArea';
+export { default as CommentBody } from './commentBody';
+export { default as CommentHead } from './commentHead';
