+++ apps/web/src/components/common/modal/index.ts
@@ -1,2 +1,1 @@
-export { default } from './authModal';
 export { default as Modal } from './modal';
