+++ apps/web/src/components/common/posting/components/index.ts
@@ -1,0 +1,1 @@
+export * from './postingHead';
