+++ apps/web/src/components/common/posting/components/postingBody.tsx
@@ -1,8 +1,8 @@
 import '@toast-ui/editor/dist/toastui-editor-viewer.css';
 
+import * as React from 'react';
 import { Container } from '@supercarmarket/ui';
 import { Viewer } from '@toast-ui/react-editor';
-import * as React from 'react';
 
 interface PostingBodyProps {
   contentHtml: string;
