+++ apps/web/src/components/inquiry/index.ts
@@ -1,6 +1,7 @@
 export { default as AdvertisementForm } from './advertisementForm';
 export { default as DealerForm } from './dealerForm';
 export { default as inquiryFormItem } from './inquiryFormItem';
+export { default as inquiryCard } from './inquiryCard';
 export { default as InquiryNavbar } from './inquiryNavbar';
 export { default as MiscForm } from './miscForm';
 export { default as PartnershipForm } from './partnershipForm';
