+++ apps/web/src/components/inquiry/inquiryCard/index.ts
@@ -1,1 +1,1 @@
-export {};
+export { default } from './inquiryCard';
