+++ apps/web/src/components/magazine/index.ts
@@ -1,4 +1,5 @@
 export { default as MagazineBanner } from './magazineBanner';
 export { default as MagazineDealer } from './magazineDealer';
 export { default as MagazineList } from './magazineList';
+export { default as MagazineCard } from './magazineCard';
 export { default as MagazineScrape } from './magazineScrape';
