+++ apps/web/src/components/magazine/magazineCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './magazineCard';
