+++ apps/web/src/components/magazine/magazineCard/magazineCard.tsx
